feat(treesitter.lua): update the list of languages to ensure installation of necessary parsers
The code in treesitter.lua has been reformatted to improve readability
and maintain consistency in indentation. Additionally, the list of
languages to ensure installation of necessary parsers has been updated
to include "dockerfile", "json", "markdown", "toml", "yaml", "python",
"http", and "html". This ensures that the required parsers for these
languages are installed and available for use with nvim-treesitter.