chore(plugins.lua): update dependency for tabline plugin to use nvim-tree/nvim-web-devicons
The dependency for the tabline plugin has been updated to use
nvim-tree/nvim-web-devicons instead of kyazdani42/nvim-web-devicons.
This change was made to align with the latest version and improvements
of the tabline plugin. Additionally, the setup function call for the
nvim-tree plugin has been removed as it is no longer necessary.