chore(neogen.lua): remove neogen plugin configuration
The neogen.lua file has been deleted, which removes the configuration
for the neogen plugin. This change was made to remove unnecessary code
and simplify the plugin setup.
The neogen plugin has been removed from the list of plugins in the
plugins.lua file. This change was made to remove unused plugins and
improve the overall performance of the application.