refactor(comment.lua): replace kommentary plugin with Comment plugin in golang and terraform configurations
The Kommentary plugin was replaced with the Comment plugin in both the
golang and terraform configurations. This change was made because the
Comment plugin provides a more streamlined and efficient way to handle
comments in the code. The pre_hook function from the
ts_context_commentstring.integrations.comment_nvim module is now used to
update the comment string, which simplifies the configuration and
improves the readability of the code.