feat(plugins.lua): add 'martineausimon/nvim-lilypond-suite' plugin for LilyPond support
The 'martineausimon/nvim-lilypond-suite' plugin is added to provide
support for LilyPond files. This plugin is configured to use 'skimpdf'
as the PDF viewer for both LilyPond and LaTeX files. An autocmd is also
added to sync syntax from the start of the buffer whenever a LilyPond
file is opened. This change enhances the functionality of the editor by
providing support for an additional file type.